[Refactor]: move httpException and config folder into common folder
