+++ src/app.module.ts
@@ -2,8 +2,8 @@ import { Module } from "@nestjs/common";
 import { AuthModule } from "./auth/auth.module";
 import { JwtModule } from "./jwt/jwt.module";
 import { ConfigModule } from "@nestjs/config";
-import { validationSchema } from "./config/config.schema";
-import { config } from "./config/config";
+import { validationSchema } from "./common/config/config.schema";
+import { config } from "./common/config/config";
 import { PrismaModule } from "./prisma/prisma.module";
 import { AcceptLanguageResolver, I18nModule, QueryResolver } from "nestjs-i18n";
 import { CompanyModule } from "./company/company.module";
