+++ src/common/config/interfaces/email-config.interface.ts
